+++ projects/journey-maps-client-testapp/src/app/services/asset-reader.service.ts
@@ -1,0 +1,14 @@
+import {Injectable} from '@angular/core';
+import {HttpClient} from '@angular/common/http';
+import {Observable} from 'rxjs';
+
+@Injectable({providedIn: 'root'})
+export class AssetReaderService {
+
+  constructor(private http: HttpClient) {
+  }
+
+  loadAssetAsString(filename): Observable<string> {
+    return this.http.get(`assets/${filename}`, {responseType: 'text'});
+  }
+}