+++ projects/journey-maps-client/src/lib/components/info-box/info-box.component.spec.ts
@@ -88,7 +88,7 @@ const createMarkerWithoutInfoBlocks = () => ({
   id: 'work',
   title: 'Office',
   position: [7.446450, 46.961409],
-  category: 'MISSED_CONNECTION',
+  category: 'WARNING',
 });
 
 const createMarkerWithInfoBlocks = () => ({