+++ projects/journey-maps-client/src/lib/model/marker-category.enum.ts
@@ -3,6 +3,6 @@
  */
 export enum MarkerCategory {
   INFORMATION = 'INFORMATION',
-  WARNING = 'MISSED_CONNECTION',
+  WARNING = 'WARNING',
   CUSTOM = 'CUSTOM'
 }